Refactor: 스타일 수정 및 목데이터 관련 리팩토링
