+++ public/data/product_nav_data.json
@@ -1,6 +1,6 @@
 [
   {
-    "id": 1,
+    "id": 0,
     "name": "dessert",
     "category": [
       { "id": 0, "name": "tart" },
