+++ src/pages/ProductList/ProductList.scss
@@ -1,0 +1,3 @@
+.productList {
+  height: 100%;
+}
